Remove GPL2+ copyright header
It seems Fatcat/cxd4 doesn't think that the plugin uses datastructures/links
against the emulator (PJ64 and M64P are both GPL2+) and disagrees with the
relicensing. He released the source originally as public domain and therefore
the relicensing is legally correct from this perspective. But keeping it not
explicitely mentioned until Fatcat changes his mind is ok for me.