Allow to set SSE optimization for builds
The default builds by the original author of the plugin were optimized for
SSSE3 or SSE2. These build options were missing in this port. Both can now be
build by setting the make parameter SSE to either SSE2 or SSSE3. It can also be
set to none when no SSE optimizations should be activated by the Makefile (own
FLAGS can still be supplied using CFLAGS).

Both build will get an own postfix in the future but currently the postfix is
used by the HLEVIDEO compile option. The postfix will be activated when the
HLEVIDEO configuration is done through the mupen64plus configuration system.